requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=9.0.0
PyYAML>=6.0
scikit-learn>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
